move some code from codeintel/lsif => codeintel/semantic
