+++ enterprise/lib/codeintel/semantic/repl/main.go
@@ -32,7 +32,7 @@ cd <path>                                             change directory
 `
 
 func main() {
-	var bundles []*conversion.GroupedBundleDataMaps
+	var bundles []*semantic.GroupedBundleDataMaps
 
 	scanner := bufio.NewScanner(os.Stdin)
 	fmt.Printf("\n> ")
@@ -139,8 +139,8 @@ func main() {
 	}
 }
 
-func queryBundle(bundle *conversion.GroupedBundleDataMaps, path string, line, character int) error {
-	results, err := conversion.Query(bundle, path, line, character)
+func queryBundle(bundle *semantic.GroupedBundleDataMaps, path string, line, character int) error {
+	results, err := semantic.Query(bundle, path, line, character)
 	if err != nil {
 		fmt.Printf("No data found at location")
 		return err
@@ -189,9 +189,10 @@ func makeExistenceFunc(path string) pathexistence.GetChildrenFunc {
 	}
 }
 
-func readBundle(dumpID int, root string) (*conversion.GroupedBundleDataMaps, error) {
+func readBundle(dumpID int, root string) (*semantic.GroupedBundleDataMaps, error) {
 	dumpPath := path.Join(root, "dump.lsif")
 	getChildrenFunc := makeExistenceFunc(root)
+
 	file, err := os.Open(dumpPath)
 	if err != nil {
 		fmt.Printf("Couldn't open file %v\n", dumpPath)
@@ -205,11 +206,11 @@ func readBundle(dumpID int, root string) (*conversion.GroupedBundleDataMaps, err
 		return nil, err
 	}
 
-	return conversion.GroupedBundleDataChansToMaps(context.Background(), bundle), nil
+	return semantic.GroupedBundleDataChansToMaps(bundle), nil
 }
 
 // finds all documents which have definition edges pointing into the argument list
-func documentsReferencing(bundle *conversion.GroupedBundleDataMaps, paths []string) (_ []string, err error) {
+func documentsReferencing(bundle *semantic.GroupedBundleDataMaps, paths []string) (_ []string, err error) {
 	pathMap := map[string]struct{}{}
 	for _, path := range paths {
 		pathMap[path] = struct{}{}
